requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
certifi==2020.12.5
chardet==3.0.4
Django==3.1.3
django-filter==2.4.0
django-ordered-model==3.4.1
django-redis==4.12.1
django-rest-auth==0.9.5
django-sequences==2.5
django-simple-history==2.12.0
django-solo==1.1.3
django-storages==1.10.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
Markdown==3.3.3
Pillow==8.0.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
